[ISSUE-187] Fix clone problem for Comp sparse vectors

+++ angel-ps/core/src/main/java/com/tencent/angel/ml/math/matrix/CSRIntMat.java
@@ -96,7 +96,7 @@ public class CSRIntMat extends TMatrix {
       }
     }
 
-    return new COOIntMatrix(row, col, rr, cc, vv);
+    return new COOIntMatrix(row, (int)col, rr, cc, vv);
   }
 
   /**
